Add code to scramble/descramble user passwords stored in the ICAT.
Recognizes non-scrambled (old) passwords and uses them unchanged.

If environment variable irodsPKey is defined for the server, use that
as the scramble/descramble key, otherwise use one defined in the
source code.  Changing this, will cause scrambled keys to no long work
right.

Passwords and be set via sql as plain-text, which will work too.  A
certain prefix string indicates when a password is scrambled.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@501 ea137bd6-aa86-47b6-b90d-b101d9f19a90