These are test scripts and input files for the iRODS TinderBox automatic build/test system.
tinderclient.pl is a major part of this, but is not included here as
I am not sure if the licensing would be OK with iRODS distributions.

We have been using various versions of these for some time (both SRB
and IRODS), but I think mantaining them in CVS will help us keep them
organized.

See the README.txt file for more.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@530 ea137bd6-aa86-47b6-b90d-b101d9f19a90