installPostgres.pl: Check on Postgres status files (/tmp/.s.PGSQL.*.lock) and allow an install even if a Postgres is running, if the ports are different.
If the old and new ports match, abort like before, with additional
explaination that the user could rerun the script and use a different
port.

irodsctl.pl:
Instead of ps (via the getProcessIds sub), use pg_ctl to get the
postgres server PIDs so that we'll only "see" "our" server PIDs.


These changes make it possible to install iRODS/ICAT when someone else
is running a postgres (installing a second postgres system).


git-svn-id: svn://irodssvn.ucsd.edu/trunk@573 ea137bd6-aa86-47b6-b90d-b101d9f19a90