icatLowLevelPostgres.c:
Removed logic that was auto-committing on any disconnect (chlClose),
and added code so that if Auditing-SQLs are pending, commit.  If only
non-Audit SQLs are pending, log some warnings messages but don't
commit.  Usual case is that nothing is pending, and so no commit is
done.  'commit' or 'rollback' empty the pending list.

icatMidLevelRoutines.c:
When using Postgres, add calls to cllCheckPending to indicate
auditing SQLs.


icatHighLevelRoutines.c:
Remove a LOG_SQL call (sql forms) that is no longer always exercised
(auto-commit).


git-svn-id: svn://irodssvn.ucsd.edu/trunk@756 ea137bd6-aa86-47b6-b90d-b101d9f19a90