finishSetup.pl: Redo a few messages to make it clear that the script is checking if the ICAT database exists.  As Dave had it, it seemed to be creating a database when we wanted to reuse one, resulting in users' heightened blood pressure...
irodssetup.pl:
Add a message that we are running 'irodsctl --quiet stop' during the
"Preparing" phase.  Dave thought it better to not mention this (avoid
confusing new users), but I believe it is more confusing to not
mention it.  For new users, it still appears as a minor message.

utils_platform.pl:
Add a fix to a recent change, redirect stderr of 'which' command since
errors are sometimes expected.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@813 ea137bd6-aa86-47b6-b90d-b101d9f19a90