Removed the option (in various places) to upgrade an existing Postgres installation, as it is confusing and dangerous and may not work, and the irodssetup script system is already doing too much.  Users should do this outside of iRODS if and when they need to.  I removed a few prompts and removed and changed some messages, but left some of the logic in the perl code for now as it's embedded in many places.
Changed message about setting system (irodsctl) to also start and stop
Postgres, to suggest they will likely want it to do so.  Changed the
prompt to default to "yes".

Removed the message about making sure that the Postgres DBMS is
running.  It is usually not true (in fact, the first step may be to
stop it).

Changed a message to have a less apologetic tone about only using
Postgres or Oracle.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@838 ea137bd6-aa86-47b6-b90d-b101d9f19a90