Fix a problem where multiple statements were being auto-committed (single statements worked OK).  This was noticed when doing 'iadmin mkuser' when group 'public' doesn't exist, resulting in an error in the rule, but the user was still being created.
In cllConnect, remove the call to SQLSetConnectOption (with
SQL_AUTOCOMMIT_OFF) as it doesn't seem to work properly with multiple
statements, and interfers with the changes below (putting a
"transaction already in progress message in the pg log).

In cllExecSqlNoResult and cllExecSqlWithResult, add a "begin"
statement call to explicitly begin a transaction, fixing the auto
commit problem.  This also prevents some error messages in the pg log:
"there is no transaction in progress".  (We sometimes still get those
errors due to multiple commits or rollbacks, but that's OK.)

While at it, cleaned up the code a little; rename cllExecSqlNoResultBV
to _cllExecSqlNoResult, make it internal, and change the arguments
some.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@861 ea137bd6-aa86-47b6-b90d-b101d9f19a90