icatLowLevelOracle.c: Add a cllConnectRda function to support RDA to Oracle (requested by Jean-Yves).
server/icat/src/icatHighLevelRoutines.c:
For the fix made to chlModAccessControl on July 10, make an change
(ifdef) so that it will work on Oracle too.  Have to cast to integer
with Oracle, bigint with Postgres.

server/icat/src/rdaHighLevelRoutines.c
Apparently, Oracle OCI does not correctly return the number of rows
available from a select (cllGetRowCount), and does not have options
to do so.  So, in rdaSqlWithResults, I added logic to handle this
case, malloc'ing a buffer and re-malloc'ing if more rows are
returned.

This Oracle-cllGetRowCount problem may affect other callers of the
general-Query too (when asking for total-rows), but this fix seems to
solve the problem for the RDA interface for when using Oracle.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@1010 ea137bd6-aa86-47b6-b90d-b101d9f19a90