m2icat.pl: Add logic to convert user-defined metadata, both dataObjects and collections, from MCAT to ICAT.  Uses 'imeta' to add the metadata.
icatHighLevelRoutines.c:
Add an internal function, _rollback, to do rollbacks and log both
success and failure (since it is doing error processing).  Change
chlRegColl to use _rollback instead of doing it in-line.

In chlAddAVUMetadata, add multiple calls to _rollback to recover after
a SQL error (such as for a duplicate entry); needed for the next SQL
to work since we now do Begin and End.  This was noticed during
m2icat.pl testing.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@1174 ea137bd6-aa86-47b6-b90d-b101d9f19a90