In igsiEstablishContextClientside, change an argument to a gsi function from 'target_name' to GSS_C_NO_NAME to fix a problem noticed by Jean-Yves (and reproduced here). I thought irods-GSI was working well (and another site tested it some to), but apparently not.
It would fail to authenticate if the GSI user was different from that
of the server, with an error about mismatched DN.  Environment
variable SERVER_DN has no effect.  It was expecting the Server's DN to
match that of the client.

More work is needed on igsi.c before the next release, as this is just
a quick patch for now, to fix the immediate problem.

The error was: "The name of the remote entity
(/C=US/O=SDSC/OU=SDSC/CN=Wayne Schroeder/UID=schroede), and the
expected name for the remote entity (/C=US/O=SDSC/OU=SDSC/CN=Wayne
Schroeder/UID=waynes) do not match".


git-svn-id: svn://irodssvn.ucsd.edu/trunk@1206 ea137bd6-aa86-47b6-b90d-b101d9f19a90