Changes to support remote zones/federation:
icatHighLevelRoutines.c:
In lots of cases, use the zone provided in the user structure rather
than defaulting to the local zone.

In chlModZone, use the input zone name instead of defaulting to the
local zone.

In chlModGroup, accept new argument userZone and use it if not null
(still default to local zone if it is).  Also, if the user is not
found and the zone is not the local zone, return success rather than
an error (so other processing can continue).

In chlRegUserRE, use the userZone if provided, otherwise default
to the local zone as before.


iadmin.c:
Change the add user command to have optional zone following, update
help to match.  Remove (for now) suggestion on @department and #zone.

Change the atg (add to group) and rfg (remove from group) commands to
allow an optional userZone and update the help to match.


icatAdminMS.c:
Add the user's zone as an argument to chlModGroup.


icatMidLevelRoutines.c:
Remove a rodslog-LOG_SQL case in cmlAudit3 as it is no longer
exercised (null zone).


icatHighLevelRoutines.h:
Add userZone as an argument to chlModGroup.


rsGeneralAdmin.c:
In the modify group command, pass another argument to chlModGroup (the
zone).


git-svn-id: svn://irodssvn.ucsd.edu/trunk@1212 ea137bd6-aa86-47b6-b90d-b101d9f19a90