icatHighLevelRoutines.c: In chlRegUserRE, chlDelUserRE chlRegCollByAdmin, and chlModGroup parse the user argument into possible user#zone and, if provided, use the zone portion.
Change a number of functions to use the user's zone in setting and
checking access permissions, etc.  Pass new zone arguments to cml
functions.

Change quite a few function calls to include new arguments (as needed
for other changes).

icatMidLevelRoutines.c:
In cmlCheckDir, cmlCheckDirId, cmlCheckDirOwn, and cmlCheckDataObjOnly
include a new userZone parameter and change the SQL for checking
access to include the user's Zone.

Change some functions and calls to functions to have an additional
bind variable argument.

icatLowLevelPostgres.c:
In cllExecSqlWithResultBV, use the 5th bind variable if provided.

icatGeneralQuery.c:
When checking access, use the a zone if provided and the localZone by
default.

icatMidLevelHelpers.c:
Change a cmlGetIntegerValueFromSql call to match a change to arguments.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@1240 ea137bd6-aa86-47b6-b90d-b101d9f19a90