Correct the code so users can specify a SERVER_DN if they want GSI to verify the identity of the server (i.e. perform mutual authentication during the GSI authentication exchange).
clientLogin.c:
In clientLoginGsi, pass in the SERVER_DN environment variable, or
NULL, as the target DN to igsiEstablishContextClientside.

igsi.c:
In igsiEstablishContextClientside, if a serviceName is passed in
(SERVER_DN), perform the gss_import_name and use the result with the
gss_init_sec_context call.  Otherwise, do as before and use
GSS_C_NO_NAME to skip the authentication of the server (just do
client authentication).


git-svn-id: svn://irodssvn.ucsd.edu/trunk@1281 ea137bd6-aa86-47b6-b90d-b101d9f19a90