clientLogin.c: For GSI authentication, use irodsServerDn if it is defined (or SERVER_DN if defined, as before).  irodsServerDn takes precedence.
Remove a debug print statement ("test1") displayed for anonymous
user.

getRodsEnv.c:
For irodsServerDn, store it into the environment (unless there is
one already) for use by clientLogin.

For the irodsAuthScheme item add a check so the environment variable
version is not overwritten if it exists.  The environment variable
should now properly take precedence, like all the rest.

igsi.c:
Don't call gss_release_name in the case where gss_import_name has not
been called.  This change seems to avoid a segfault when SERVER_DN is
not defined.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@1478 ea137bd6-aa86-47b6-b90d-b101d9f19a90