During the GSI authentication sequence (after the DN is confirmed), if there was no irodsUserName provided:
1) Query on just the DN and set rsComm->clientUser.userName to the
userName found if one and only one matching DN is found.

2) If that fails, run the rule acGetUserByDN which is a no-op by
default but can be configured to run a command that returns (in
stdout) the irods user to use.  This will support VOs.

3) If that doen't return a result, do the DN query again as the rule
might have added the user.  This makes it possible to automatically
add irods users when GSI users first connect.

In each of these 3, if it succeeds, set rsComm->clientUser.userName to
the returned string.

This was done in collaboration with Shunde Zhang and Graham Jenkins of
the Australian Research Collaboration Services (ARCS).


git-svn-id: svn://irodssvn.ucsd.edu/trunk@1484 ea137bd6-aa86-47b6-b90d-b101d9f19a90