Modifications from Andy Sainikov (SLAC) to support MySQL as the ICAT DBMS (when previously-installed).  This was successfully tested here at CA-DICE on Ubuntu and seems to work fine.  Other testing is on-going and we will probably want to extend the installation support and documentation and make some other refinements.
The install scripts now include options for MySQL, altho Andy's
procedure does not include irodssetup but instead some of the
sub-scripts.  See INSTALL.txt for instructions.

With these mods, the icatLowLevelPostgres.c code now works with both
PostgreSQL and MySQL so we may rename it sometime.

The two new .pp files are now run thru cpp to generate MySQL,
Postgres, or Oracle versions (with slight differences), a nice way to
handle this.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@1523 ea137bd6-aa86-47b6-b90d-b101d9f19a90