Add a 'loadtest' command which runs the concurrent poundtest suite, showing the steps as they complete.
In 'loadtest', by default create 100 MB bigfiles to trigger parallel
threads but show code to change to smaller ones.

Fix a couple problems so can run 'devtest' as well as 'loadtest' on a
client-only machine: Improve the check for the IES being local and
prompt for the password if needed (for icommands test).

Change the 'test' command to 'devtest', developer test.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@1651 ea137bd6-aa86-47b6-b90d-b101d9f19a90