Add logic to better handle GSI authentication when the user's name is not specified.  Added logic so it doesn't rely on the client-provided zone name, since it may not be specified or could be the wrong zone. Instead get the zone from the ICAT and store it in the appropriate items in rsComm.  In the case where the rule is running an external process, parse the results into username and zone and use those if of that form (user#name) (this part is untested tho).
This should fix a problem noticed by Shunde and Lucas where if the
zone is not specified on the client side, the login might succeed and
queries might work but other operations, like put, would fail.  The
problem was that the user's rodsZone in the rsComm was null, creating
invalid SQL for some access checks.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@1698 ea137bd6-aa86-47b6-b90d-b101d9f19a90