rsGeneralAdmin.c: Call the new replErrorStack to return the error stack if rcGeneralAdmin fails.  This returns the error stack messages if the user is connecting thru a non-ICAT-Enabled server.
This is primarily for the new CAT_PASSWORD_ENCODING_ERROR which will
happen if the admin-user has connected thru a non-ICAT server.  So
this will give them a more detailed message about what is wrong and
how to correct it.

But this also helps for other admin functions, if the user happens to
be connecting thru a non-IES since the iadmin interaction is designed
to use the error-stack for additional assistance.

Also made a minor logging change.


icatHighLevelRoutines.c:

Minor refinement to the error-stack text message returned for the case
of CAT_PASSWORD_ENCODING_ERROR.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@1814 ea137bd6-aa86-47b6-b90d-b101d9f19a90