server/icat/src/icatGeneralQuery.c:
When forming the where clause, if the condition is prefixed with 'n',
add SQL to cast the column to decimal (which is equivalant to numeric
and works for MySQL too).  This will be used by imeta and to query in
numeric mode on AVUs and will be available to other clients too.

clients/icommands/src/imeta.c:

Add to the help text for the query to explain the new numeric mode
(prefix with 'n').  No execution code changes were needed as this is
all handled on the ICAT side.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@1951 ea137bd6-aa86-47b6-b90d-b101d9f19a90