Add a new index on the main collection table, needed for the ICAT to scale properly as the number of collections increases.  This was recently determined as a key performance issue for a site with about 250,000 collections.
In local ICAT-intensive tests, some operations that had slowed by more
than an order of magnitude under these conditions returned to nearly
the original speed.  For example, on modest hardware, an iput -r of
500 small files took about 9 to 10 seconds with just a few defined
collections without this index, took 350 seconds (and sometimes much
more) with 250,000 collections without this index, and again took
about 10 seconds with the index.

The name of the patch file will be changed depending on what we call
the next release.  We will also need patch files for Oracle and MySQL.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@2170 ea137bd6-aa86-47b6-b90d-b101d9f19a90