Added 'groupadmin' capability.  Users of type 'groupadmin' can add and remove other users to/from groups they are members of themselves.
iadmin.c:
Add an option to the iadmin.c generalAdmin function to optionally also
try rcUserAdmin, if rcGeneralAdmin fails with permission error.
Invoke this option for 'atg' and 'rfg'.

Add help text in 'atg' and 'rfg' for 'groupadmin'.


rsUserAdmin.c:
Add "modify" "group" option similar to the one in rsGeneralAdmin.
This calls chlModGroup which checks acess (groupadmin).

Made some cosmetic code changes to make it a little more readable, at
least for me.

icatHighLevelRoutines.c:
In chlModGroup, if the caller is not LOCAL_PRIV_USER_AUTH (rodsadmin),
check if the user is 'rodsgroup' and a member of the group by calling
cmlCheckGroupAdminAccess.

icatMidLevelRoutines.c:
Add function cmlCheckGroupAdminAccess.

icatMidLevelRoutines.h:
Add definition for new cmlCheckGroupAdminAccess.

icatTest.pl:
Add a test to exercise the new SQL code for the 'groupadmin'
capability.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@2204 ea137bd6-aa86-47b6-b90d-b101d9f19a90