Add to the scripts so that 'irodssetup' can handle the case where the user is installing Postgres from scratch and wants to use a Database admin user name that is different from the OS username.  (In other cases, it is up to the user to create the DB user account before running 'irodssetup').
finishSetup.pl: 
Add a function create the new Postgres user, if it is not the OS
username, and call it before switching to using that user name.  
Also, change the way .pgpass is set up so it will be useful in either
case (DB user same as OS, or not).

utils_config.pl:
Add 'createuser' with the full path for use in finishSetup.pl.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@2249 ea137bd6-aa86-47b6-b90d-b101d9f19a90