server/icat/src/icatGeneralQuery.c:
Fix a couple bugs dealing with 'cycler's (which have been re-added for
quotas): In generateSQL, when calling tScan, be sure to start with a
non-cycler.  In tCycleChk, handle cycler tables in similar manner as
the actual tScan does (return).


server/test/src/test_genq.c:

Make the results of a full table scan test a litte clearer and exit
based on the result.


server/test/bin/icatMiscTest.pl:

Add a call to the scan test (check for cycles in the table links) so
it will be done automatically via tinderbox.


server/icat/src/icatGeneralQuerySetup.c:

Fix some problems with the new quota tables: Make r_quota_main and
r_quota_usage non-cyclers so that can link with the related tables.
Correct the names of two columns.  Add links for the
r_quota_user_group table.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@2279 ea137bd6-aa86-47b6-b90d-b101d9f19a90