In addInClauseToWhere (an optimization added to 2.2, provided by Andy), insert the substrings for the IN clause as bind variables instead of quoted strings, in case there are single or double quotes in the string (file path name).
This was noticed by Pauline Mak of ARCS (and discussed on irods-chat),
and would cause iput failures for a files with quotes and spaces (such
as 'How I PC'd an Apple G5".rtfd.zip').  The bug caused a SQL error
which seems to cause an iput failure under 2.2.  In current SVN code,
it was causing the SQL error but no iput failure (apparently due to
changes in caller code (rsObjStat.c, I believe)).



git-svn-id: svn://irodssvn.ucsd.edu/trunk@2283 ea137bd6-aa86-47b6-b90d-b101d9f19a90