iTestGenQuery.c: Add a test based on a function that Jean-Yves is using for the i-command he is developing.  When this function is called many times (80 to 90 or so) in a single session (one agent), it can trigger an ICAT bug, either a seg-fault or some other error (in my case, a CAT_INVALID_ARGUMENT).
icatGeneralQuery.c:
Fix the above problem.  Need to reinitialize the static int
conditionsForBindIx in the handleCompoundCondition function so that
the scratch buffer space it is using won't overflow.  Problem would
only appear when large numbers of GenQuery calls were made in a single
session and those queries included compound conditions.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@2354 ea137bd6-aa86-47b6-b90d-b101d9f19a90