Fixes for bugzilla bug 53 and 54, and related changes.
In many cases, changed the size argument for some snprintf, strncpy,
strncat, and rstrcpy calls to use 'sizeof' the target instead of
defines or matching constants.  In almost all of these, the previous
form was correct but the sizeof form is clearer and less error prone.
The rare cases where these values were wrong, as described in bugs 53
and 54, are corrected in this set of changes (in rsGeneralAdmin.c and
some of icatHighLevelRoutines.c respectively).

I checked many other cases of these types of calls too but did not
change them as they appear to be correct and clear and there are so
many (in my code and others').

But I do recommend the sizeof form in new code and will tend to use
it.  I may also modify other sections of source in this manner when I
am working on them for other reasons.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@2363 ea137bd6-aa86-47b6-b90d-b101d9f19a90