Add code for a quota-query to be used by the server when enforcing quotas, and make some minor refinements in the quota system.
icatGeneralQuery.c: 
Add code to handle a specific query, when the new QUOTA_QUERY option
is used, to efficiently get quota information, handling all the quota
types in a single query (the group and individual quotas, per-resource
and total-usage).  I decided to make this part of General-Query, since
it is so similar, but it is specific.  See the comments and
iTestGenQuery.c for more.

iTestGenQuery.c:
Add a test/example of the new QUOTA_QUERY for getting quota values.

rodsGenQuery.h:
Add special QUOTA_QUERY option.

icatHighLevelRoutines.c:
Re-enable the call to setOverQuota when setting a quota; I believe it
will maintain consistency better and should always be quick.
Do a little clean-up.

quotaTest.pl:
Adjust the test to match the above changes in icatHighLevelRoutines.c;
'iadmin cu' is not needed before the new values are seen.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@2380 ea137bd6-aa86-47b6-b90d-b101d9f19a90