For quotas, where it wasn't already handled, add logic to deal with optional zone names in user names.  As usual, default to the local zone.
iquota.c:
Parse out the optional zone on a username and use it, or the local
zone, when doing queries.  Display zones as part of the output.
Adjust the help text to match.

iadmin.c:
Change the simpleQuery SQL for 'lq' (list quotas) to include a zone
and use either the entered zone or the localzone.  Also, include the
zone in the display.  Update the 'lq' help text to match.

icatHighLevelRoutines.c:
Change the chlSimpleQuery SQL for quotas to include user zones.

icatGeneralQuery.c:
In the special query for quota enforcement, parse out an optional zone
on the user name and use it, or the local zone.

icatGeneralQuerySetup.c:
Set up the new quota user-zone column.

rodsGenQuery.h and rodsGenQueryNames.h:
Add a new column for general queries on quotas for user-zones.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@2422 ea137bd6-aa86-47b6-b90d-b101d9f19a90