Fix a couple quota bugs.
In setQuota, when handling group qoutas on resources, change the SQL
to also get the resource id and then use that when setting the values.
The quotaTest.pl does catch this problem, so I'm pretty sure it was
working properly perviously so another change seems to have caused
this.  But this change seems to correct the problem and quotaTest.pl
now succeeds.

Also, change the way group quotas for total-usage is handled to avoid
setting the over-quota too high in some cases.  Do it in separate SQL
calls and only set the over-quota value when it is higher than
previous settings (based on per-resource quotas).  This was a bug that
Mike noticed in some of his testing.

Also did a little code cleanup, including a simplier way to handle
'cast' differences between the 3 DBMS types.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@2434 ea137bd6-aa86-47b6-b90d-b101d9f19a90