Avoid a connection failure in a GSI case noticed by Shunde.
If using GSI, and if the client does not provide a user name, and is
connecting to a non-IES, avoid a particular query (which would fail)
and set the privilege level to local user.  This avoids a connection
failure (and non-IES abort) in this case, but limits the users to
non-admin operations.  It is still secure because the non-IES is
authenticated and then verifies the client user via GSI.  Extensions
could be made to the iRODS protocol, after the upcoming release (2.3),
to also handle the admin case, via a new privileged client-server call
to adjust the client user's privilege level.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@2510 ea137bd6-aa86-47b6-b90d-b101d9f19a90