Add and update tables for holding the rules in the ICAT.
Update the columns in R_RULE_MAIN to match the current definitions
(like those in the 2.2 to 2.3 patch files), and add the R_RULE_DVM and
R_RULE_FNM (as per the patch files too).  I had forgotten to go back
to the icatSysTables.sql and make these more recent changes (since
these tables were not used in 2.3).

We will need to account for this for the patch files for the next
release, as sites who ran the patch will have the right tables but
those who installed 2.3 from scratch will need to have them dropped
and recreated.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@2601 ea137bd6-aa86-47b6-b90d-b101d9f19a90