Add the capability of renaming a resource.
icatHighLevelRoutines.c:
Add logic to handle renaming a resource, updating the various tables
that include a resc_name column.

iadmin.c:
Warn and prompt for confirmation when renaming a resource.
Add 'name' to modresc description in the help text.

icatTest.pl:
Add a test for changing the name of a resource so that all the SQL
forms will be tested.

icatGeneralQuery.c:
Correct a log message.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@2622 ea137bd6-aa86-47b6-b90d-b101d9f19a90