Fix for GSI authentication, connecting to non-IES, without irodsUserName.
Some sites that use GSI authentication find it useful to dynamically
map the GSI authentication id (DNs) to irodsUserNames rather than
having the user specifying them or by maintaining a separate map file.
There is a Rule that can be used for this, or if the DN is associated
with exactly one irods user, the iRODS code will use that.  See
server/api/src/rsGsiAuthRequest.c for more information on this.

This has been working fine in previous versions, except when the
client connects to an non ICAT-Enabled Server (non-IES) instead of the
IES.  With these changes, connecting to the non-IES using GSI without
specifying an irodsUserName works too.

This problem was reported by Shunde (ARCS).


icatHighLevelRoutines.c:
In chlCheckAuth, if the rodsUserName is blank, instead of returning an
error, allow basic access (the remote server has been authenticated so
we can trust this).  This will allow some queries from the non-IES,
including the one needed to get the DNs from the ICAT.


rsGsiAuthRequest.c:
If the rodsUserName was not previously set, but is now (via various
methods), and this agent is not the ICAT-Enabled Server (IES),
disconnect from the IES and reconnect.  This adds a little overhead,
but cleanly and securely sets up the environment in the IES agent
process for this user (and some sort of additional interaction is
needed anyway).


git-svn-id: svn://irodssvn.ucsd.edu/trunk@2639 ea137bd6-aa86-47b6-b90d-b101d9f19a90