Add the 'imeta addw' wildcard selection of data-objects capability. This was more involved than I expected (particularly for the access checks) but seems to work well and should scale well.  Users can create an AVU and associate it with many data-objects at one time, using wildcards in the collection and/or data-object name.
Like similar operations, this is available for use from microservices
too.  Note that the return code, if positive, is the number of objects
to which the AVU was associated.

icatHighLevelRoutines.c:
Add the chlAddAVUMetadataWild function (see comments for descriptions).
Refactor chlAddAVUMetadata to have a separate function, findOrInsertAVU, 
which is used by both chlAddAVUMetadata and chlAddAVUMetadataWild.

imeta.c:
Add the 'addw' sub-command and its help informtion.

rsModAVUMetadata.c:
Handle the new 'addw' option; call the new
chlAddAVUMetadataWild function.

icatTest.pl:
Add a basic test of 'imeta addw' so that the SQL will be exercised and
the tinderbox test ('irodsctl devtest') can succeed.

icatDefines.h:
Add a new Audit type for use with chlAddAVUMetadataWild.

icatHighLevelRoutines.h:
Add a definition for the new chlAddAVUMetadataWild function


git-svn-id: svn://irodssvn.ucsd.edu/trunk@2680 ea137bd6-aa86-47b6-b90d-b101d9f19a90