Changes for the acGetUserByDN authentication mode developed in collaboration with ARCS/NZ (Vladimir Mencl).
This is the mode where users authenticate via GSI, without specifying
the irodsUserName, and the acGetUserByDN rule is configured to run an
external script to create the user.

The changes correct problems when running this mode on a non-IES and
set up the environment as expected by the ARCS script that is run via
acGetUserByDn.  Instead of possibly using the stdout from the script,
the function now always re-does the query on DN to verify the new
user.  Also, the SP_CLIENT_USER (spClientUser) environment variable is
reset with the new username in this situation as the script makes use
of this.* A test has been modified to avoid a seg-fault in a error
case.

* SP_CLIENT_USER is set to the irodsUserName and so in this mode is
initially null.  The default commands that can be executed (in
server/bin/cmd) do not make use of this, but the ARCS script does.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@2713 ea137bd6-aa86-47b6-b90d-b101d9f19a90