icatHighLevelRoutines.c: For Database-Objects use some specific AVUs to hold DBO information (instead of the data-object comment column and perhaps the path) and instead of creating a special data-object as a DBO, just assign DBOness to data-objects.  This avoids reusing columns for multiple purposes, uses existing AVU access control, and allows the root object to be a normal data-object.
And with that, we could store the results of the SQL (select) into the
data-object, which might be quite useful.  We could have separate
control to execute the SQL and then use normal iRODS access control
for other users to access the results.  Longer term, we may be able to
take this data stored in the data-object and insert it into another
table.

idbo.c:
Optionally use the current-working-directory when specifying
data-objects.

rsDatabaseObjectAdmin.c:
Remove some debug printf's.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@2941 ea137bd6-aa86-47b6-b90d-b101d9f19a90