Consistently use uppercase names for tables, for MySQL, fix bug 22.
In much of the code, we would use upper case names for the ICAT tables
but in many cases we'd use lower.  This is fine for Postgres and Oracle and
it works for MySQL when configured with	lower_case_table_names=1.

With these changes, one can also run mySQL in the regular
case-sensitive mode as requested in Bugzilla item to 22.  The tables
are installed in upper case but the SQL consistently matches it.

In testing, I found that if I tried to alter the MySQL ICAT tables to
a different case, rows were lost (perhaps those added as the other
case name).  So I am not sure this will work for existing MySQL ICATs
(i.e. sites may need to continue to run with lower_case_table_names=1).  
But for new installations, this seems to work.  At least the
installation succeeds and the full ICAT test ('irodsctl devtest')
completes successfully.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@3006 ea137bd6-aa86-47b6-b90d-b101d9f19a90