Adjust chlSimpleQuery to old iadmin will work with new servers.
Change the check on the input SQL to chlSimpleQuery 
be case insensitive so that older versions iadmin will still be
able to operate with the server even with the changes committed
earlier today in which table names are always upper case.  The
old iadmin inputs table names to chlSimpleQueyr in lower case.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@3008 ea137bd6-aa86-47b6-b90d-b101d9f19a90