Add a --no-page option to iquest to bypass paging prompt.
Add the --no-page as a standard long option in the command line
parsing function/structure and make use of it in iquest.  When
specified, bypass the default paging.

This was requested by two groups (Graham Jenkins and then Scott
Sinno).  The paging feature was originally added at the request of
Florian Goessmann and it does seem like a good way to handle
unexpectedly long set of responses to a query (prompt and ask if you
want to continue after a large number (500 currently)).  But if you
know the results might be long and you don't want to page like that,
you can include --no-page on the command line.  You can pipe 'yes' to
'iquest' but you'd still could have the prompt strings in the output
to deal with.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@3094 ea137bd6-aa86-47b6-b90d-b101d9f19a90