Six more SQL indexes that should improve performance in some cases.
We had an initial set of indexes from the original schema design, and
added a few in response to particular performance issues that arose,
but now we (Raja) went thru and looked at what could be added or
removed.  There are other possibilities too which we will investigate
in more detail later but we wanted to get these into the 2.4 release.

These were


git-svn-id: svn://irodssvn.ucsd.edu/trunk@3106 ea137bd6-aa86-47b6-b90d-b101d9f19a90