When updating LOAD_LIBRARY_PATH, append /usr/local/lib instead of prepending it.
This was an issue found by John Knutson (U of Texas) and described on
irods-chat.  There was a difficult bug to resolve because the wrong
version of HDF libraries were being loaded from /usr/local/lib.

This should correct the issue in this case and perhaps others.
Generally, it should be better to append /usr/local/lib so that local
customizations to LOAD_LIBRARY_PATH take precedence.

The database libraries are still prepended since they are
build-specific.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@3291 ea137bd6-aa86-47b6-b90d-b101d9f19a90