Return an error if multiple tests on Collection AVUs are requested (COL_META_COLL_ATTR_NAME) as this is currently unimplemented.
Without this addition, the general-query would return inaccurate
results in this situation, effectively ignoring the additional
conditions.  If queries like this are needed, we can add code to
handle them, to adjust the generated SQL (something similar to the
'handleMultiDataAVUConditions'). 'imeta' help text explains that
multiple AVUs tests are available for Data AVUs (not Collections) but
the server side (this function) needs to catch this for all clients
(both imeta and Jargon).

There is a second issue with 'imeta', as it ignores the trailing
second condition.  So even with this change, multiple tests won't
work right.  For example:
  imeta qu -C m1 = 1 and m2 = 2
will come into the server as if the user entered:
  imeta qu -C m1 = 1



git-svn-id: svn://irodssvn.ucsd.edu/trunk@3492 ea137bd6-aa86-47b6-b90d-b101d9f19a90