Example DBOs that also serve to list tables within a DBR.
With recent changes to how we plan to implement DBRs/DBOs, it is now
easy to implement these list-tables DBR operations as additional DBOs
(instead of a built-in SQL like I had it).  If a site wants to provide
this, they can store the DBO with appropriate attributes and users
will be able to simply execute the DBO on the DBR via 'idbo'.  Some of
this requires some changes to DBO functions not yet committed.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@3523 ea137bd6-aa86-47b6-b90d-b101d9f19a90