This is the initial implementation of the SQL-based queries, a.k.a. the specific-query, the ability to pre-define some SQL queries that users can execute.
Care must be taken when using these as they by-pass the access
controls of the general-query and require detailed knowledge of the
ICAT schema (where as the general-query does not).  For example, when
there is a general-query on data-objects or collections, the various
modes of access control are enforced via additional SQL, where as the
specific-query will only do whatever the admin tells it to.  Another
example is that the user passwords are not accessible at all via the
general-query, but could easily be displayed via a specific query if
the admin inadvertently allows it.

But the specific-query may be useful in some cases where a site wants
to execute some specific SQL to handle a particular case, working
around some limitations of the general-query.

The approach is as follows:

1) There is new iadmin command to add and remove the allowed sql-query
SQL strings, 'asq' and 'rsq', add and remove specific query.  See
'iadmin h asq'.  The help text warns the admin to be careful about
adding approved SQL strings as normal access controls are bypassed.

2) These SQL strings are stored in a new ICAT table.
server/icat/src/icatSysInserts.sql and icatSysTables.sql have been
updated to create the new ICAT table, r_specific_query, and to set up
the two built-in queries (to list the queries).

3) The caller sends the SQL string and parameters (if any), and the
server ICAT code makes sure the SQL string is in the allowed set.  It
does this via a query on this new table (this makes each query do two
queries, but the first one is trivial and so very fast).

4) One of the sql-queries (preset) is one that returns all the allowed
specific-query SQL strings. (see 'iquest --sql ls').

5) Another one does some string matching on the SQL strings just a SQL
doing a 'like' on a string (user can input wildcards (e.g. '%demo%')).

6) The SQL form may include parameters as bind variables and this is
in the Postgres format (using ? for each) and the ICAT code converts
this to the Oracle format for Oracle ICATs.

7) The results are returned in the same type of structure as the
general-query, altho the attribute index is not stored since the
specific-query ICAT code does not know them in this case.

8) There is a new client-server ('api') call for this,
rcSpecificQuery.

9) An option to the iquest i-command has been added to support this, a
new command line option, --sql.  See  'iquest h'.

This commit also adds some tests to icatTest.pl so that the new SQL
forms will be tested and the 'tinderbox' runs can succeed.

There is also a new error code.

It would be useful to add an option to use a format statement in
'iquest --sql' so that the returned values could be formatted in a
flexible manner.  But currently, it's just '--sql', the 'SQL
statement' and, if needed, bind-variable arguments.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@3703 ea137bd6-aa86-47b6-b90d-b101d9f19a90