Minor update for additional log information.
With the previous version, r4050, the log information on
icatHighLevelRoutines.c was accidentially destroyed (Mike
inadvertantly did an 'svn mv').  Also, that version, r4050, actually
had no changes from the previous version but was commmitted because of
the accidental 'svn mv'.

But log information and comparisons with previous revisions are still
possible.  You can run 'svn update -r r4049 icatHighLevelRoutines.c'
and then a 'svn log icatHighLevelRoutines.c' to see those messages.
Mike also added the file icatHighLevelRoutines.c.log for to record the
svn log messages for the updates on icatHighLevelRoutines.c prior to
r4050.

Comparisons can be done in a similar manner, after first doing the
svn update to an old revision. For example:
  svn update -r r4049 icatHighLevelRoutines.c
  svn diff -r r3915 icatHighLevelRoutines.c



git-svn-id: svn://irodssvn.ucsd.edu/trunk@4085 ea137bd6-aa86-47b6-b90d-b101d9f19a90